requirements.txt
alembic==1.4.2
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
get==0.0.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
post==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
query-string==0.0.0
request==0.0.0
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.18
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
